Spawn traders at tradeposts, and make them cycle around.
Trader.prototype.PerformTrade and UnitAI.prototype.PerformTradeAndMoveToNextMarket
only have a small change that returns the next market in the former and
uses that in the latter to enforce information hiding.
(This part should be merged upstream.)

The other functions in the components allow creating a trade road and
should be used as a foundation for #3872. It is missing some error
checking in the UnitAI function, as it expects to be called with only
valid parameters and targets, which we do have in this map. Whether
formations work was not checked, but should. Something to think about
would be how to actually expose that to the user without breaking the
lazy trade route setup (which waits for two markets to be specified,
which is not working with this code path, and it would be nice if we
could merge both of these).